refactor: getting skill categories
