+++ src/hooks/useGetUsers.ts
@@ -1,17 +1,17 @@
-import { GET_SKILLS_QUERY_KEY, getSkills } from '@/services/skill.service';
 import { GET_ALL_USERS, getUsers } from '@/services/user.service';
-import { Skill } from '@/types/skill.type';
 import { User, UserWithPercent } from '@/types/user.type';
 import { matchingIndicator } from '@/utils/matchingIndicator';
 
 import React, { useRef } from 'react';
 
 import useAuth from './useAuth';
+import useSkill from './useSkill';
 
 import { useQuery } from '@tanstack/react-query';
 
 const useGetUsers = () => {
   const { user } = useAuth();
+  const { skillMapping } = useSkill();
   const matchCache = useRef(new Map<string, UserWithPercent>());
 
   const { data, isLoading, isError } = useQuery({
@@ -19,32 +19,6 @@ const useGetUsers = () => {
     queryFn: () => getUsers(user?.id || ''),
     enabled: !!user?.id,
   });
-
-  const { data: skills } = useQuery({
-    queryKey: [GET_SKILLS_QUERY_KEY],
-    queryFn: () => getSkills(),
-  });
-
-  const skillsMap =
-    skills?.reduce(
-      (map, skill) => {
-        map[skill.id] = skill;
-        return map;
-      },
-      {} as Record<string, Skill>,
-    ) || {};
-
-  const skillMapping = React.useCallback(
-    (user: User) => {
-      if (!skills || !Array.isArray(user.learn) || !Array.isArray(user.teach)) {
-        return { learn: [], teach: [] };
-      }
-      const learning = user.learn.map((id: string) => skillsMap[id]?.name || 'Unknown Skill');
-      const teaching = user.teach.map((id: string) => skillsMap[id]?.name || 'Unknown Skill');
-      return { learning, teaching };
-    },
-    [data],
-  );
 
   const processUsers = React.useCallback(
     (users: User[]) => {
